requirements.txt
appdirs==1.4.3
dj-database-url==0.4.1
Django==1.9.4
packaging==16.8
Pillow==3.3.0
psycopg2==2.6.1
pycrypto==2.6.1
PyMySQL==0.7.6
pyparsing==2.2.0
six==1.10.0
waitress==0.9.0
whitenoise==3.2
